nrf binpack file extension cpy in region 4
The .cpy file extension indicates a binary FAT filesystem image
containing a circuit python script. This must be encoded as region 4 so
that the Notecard knows to transmit the binary to the nRF host using a
protocol specific to programming the QSPI flash (as opposed to the nRF
internal Flash).