Total binary length in binpack header of manifest
The Adafruit / Nordic DFU protocol requires the binary length in the
DFU_START_PACKET, which must be sent prior to sending the metadata
(.dat) file to the bootloader.  Therefore place the binary length in the
manifest information since the address is otherwise unused and the
manifest is first in the binpacked data.