NF-477: Fix some DFU issues. (#22)
1. This undoes the earlier change to use a cmd instead of a req on the last
dfu.put of a sideload. This makes the CLI behave as it did before, where it'll
wait for a reply to dfu.put with "pending":false so it knows the DFU is "ready."
However, this once again breaks sideload for NOTE-LORA, which doesn't respond to
the last dfu.put. We are going to fix the LORA firmware so that it DOES respond
to the last dfu.put, rather than adding special logic to handle this case in the
CLI.
2. Sideload now won't error out if it sees a "firmware update is in progress"
error in response to dfu.put while waiting for pending to be false. Similarly,
if it sees the error "{dfu-in-progress}", it won't error out.

Original revert message of the earlier commit retained below:

Revert "work around the fact that old lora firmware doesn't send {} during last dfu.put"

This reverts commit 131a4eea0daac6b48b9e61094399a67a270ef935.