Make sure to close Notecard connection whenever CLI exits. (#28)
I fixed this for `-play`, but I noticed it wasn't fixed for `-trace`. I also
noticed that it wasn't closing the connection properly on signals
(e.g. SIGTERM). I reworked the code a bit so that now we make sure to call close
when a signal terminates the CLI AND in other, non-signal-triggered exit
scenarios by using `defer`.