Notecard CLI: Return from main rather than doing os.Exit(0). (#29)
This ensures the deferred cleanup function that closes the Notecard gets called.
Doing os.Exit(0) will not call deferred functions. The problem is that if you do
this:

```
notecard -port /tmp/ns_card_usb -req '{"req":"card.version"}'
```

And run it again after that, the second time will say the port is still busy,
because it was never properly closed.