NF-1336: Fix some sideload bugs. (#48)
Prior to this commit, when sideloading, we were only putting the Notecard into
"DFU op mode" for host DFU and not Notecard DFU. When you do {"req":"hub.set",
"mode":"dfu"}, the Notecard resets a 15 minute timer during which the sideload
 is allowed to take place. So, the bug I was seeing was that I had sideloaded
 Notecard firmware once, and more than 15 minutes later I tried again, but I
 was blocked because this timer had expired from the first DFU, and the CLI was
 not doing the hub.set to reset the timer. This commit fixes this problem so
 that we always do a hub.set mode:dfu at the start of a sideload and a hub.set
 mode:dfu-completed at the end. BUT, LORA doesn't support this op mode, so we
 don't do it for LORA.

The other bug was in loadBin. There, at the end of each iteration of the main
sideloading loop, we were waiting for dfu.put to return pending:false, which
indicates that the migration of the chunk from internal to external storage has
completed. However, for Notecard firmware sideload, there is no migration. The
data either flows from the binary storage region or from the payload in the
dfu.put directly into the Notecard MCU's flash. So there is no need to wait for
pending:false. I had somehow gotten a Notecard into a state where it thought it
had migration work to do, and that caught me in an infinite loop when I tried
to sideload because pending was always coming back true.

I tested this on ESP, LORA, r5, and u5 SKUs. Also tested ODFU with Swan.